// src/modules/invoice/usecase/find-invoice/find-invoice.dto.ts
import { InvoiceItemEntity } from "../../domain/invoice-items";

export interface FindInvoiceUseCaseInputDTO {
  id: string;
}

export interface FindInvoiceUseCaseOutputDTO {
  id: string;
  name: string;
  document: string;
  address: {
    street: string;
    number: string;
    complement: string;
    city: string;
    state: string;
    zipCode: string;
  };
  items: InvoiceItemEntity[];
  total: number;
  createdAt: Date;
}